Allow input, output while reading rule from file
I'm not sure this is the intended behaviour but I have a static rule body without `INPUT` or `OUTPUT` and wish to supply this dynamically. Before this commit, it was either possible to 1) read everything from file or 2) supply everything (having to pass a body in addition to I/O).