Include resc_hier in replica information
Finding out the full resource hierarchy is important for dealing with replicas. In the past, we had a naming scheme that allowed us to work around it, but now that naming scheme doesn't provide what we need. Including resc_hier in the replicas information allows us to do this the right way